backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pydantic-settings
python-dotenv
python-telegram-bot
APScheduler
sse-starlette
bleach
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
httpx
bcrypt<4
alembic
